+++ onnxruntime/core/providers/dml/OperatorAuthorHelper/OperatorHelper.cpp
@@ -1888,7 +1888,7 @@ namespace OperatorHelper
         m_outputShape.resize(2 + m_imageShape.size());
         m_outputShape[0] = m_inputShape[0];                     // N
         m_outputShape[1] = m_inputShape[1] / blockShapeProduct; // C
-        for (int i = 2; i < m_outputShape.size(); i++)
+        for (size_t i = 2; i < m_outputShape.size(); i++)
         {
             m_outputShape[i] = m_imageShape[i - 2];
         };
